feat: traits for metadata and woo
